Siva Reddy | BAH-2464 | Refactoring patient search code to bahmni-commons
