+++ bahmnicore-api/src/main/java/org/bahmni/module/bahmnicore/dao/PatientDao.java
@@ -1,7 +1,7 @@
 package org.bahmni.module.bahmnicore.dao;
 
-import org.bahmni.module.bahmnicore.contract.patient.PatientSearchParameters;
-import org.bahmni.module.bahmnicore.contract.patient.response.PatientResponse;
+import org.bahmni.module.bahmnicommons.contract.patient.PatientSearchParameters;
+import org.bahmni.module.bahmnicommons.contract.patient.response.PatientResponse;
 import org.openmrs.Location;
 import org.openmrs.Patient;
 import org.openmrs.RelationshipType;
@@ -10,15 +10,6 @@ import java.util.List;
 import java.util.function.Supplier;
 
 public interface PatientDao {
-
-    /***
-     * Please do not use this method, use the getPatients(PatientSearchParameters ...) instead.
-     */
-    @Deprecated
-    public List<PatientResponse> getPatients(String identifier, String name, String customAttribute,
-                                             String addressFieldName, String addressFieldValue, Integer length, Integer offset,
-                                             String[] patientAttributes, String programAttribute, String programAttributeField,
-                                             String[] addressSearchResultFields, String[] patientSearchResultFields, String loginLocationUuid, Boolean filterPatientsByLocation, Boolean filterOnAllIdentifiers);
 
     List<PatientResponse> getPatientsUsingLuceneSearch(String identifier, String name, String customAttribute,
                                                        String addressFieldName, String addressFieldValue, Integer length,
