+++ bahmnicore-api/src/main/java/org/bahmni/module/bahmnicore/service/BahmniPatientService.java
@@ -1,8 +1,8 @@
 package org.bahmni.module.bahmnicore.service;
 
-import org.bahmni.module.bahmnicore.contract.patient.PatientSearchParameters;
+import org.bahmni.module.bahmnicommons.contract.patient.PatientSearchParameters;
 import org.bahmni.module.bahmnicore.contract.patient.response.PatientConfigResponse;
-import org.bahmni.module.bahmnicore.contract.patient.response.PatientResponse;
+import org.bahmni.module.bahmnicommons.contract.patient.response.PatientResponse;
 import org.openmrs.Patient;
 import org.openmrs.RelationshipType;
 import org.openmrs.annotation.Authorized;
@@ -11,9 +11,6 @@ import java.util.List;
 
 public interface BahmniPatientService {
     public PatientConfigResponse getConfig();
-
-    @Authorized({"Get Patients"})
-    public List<PatientResponse> search(PatientSearchParameters searchParameters);
 
     @Authorized({"Get Patients"})
     List<PatientResponse> luceneSearch(PatientSearchParameters searchParameters);
