+++ bahmnicore-omod/src/main/resources/config.xml
@@ -32,6 +32,7 @@
         <require_module>org.openmrs.module.rulesengine</require_module>
         <require_module>org.openmrs.module.episodes</require_module>
         <require_module>org.openmrs.module.auditlog</require_module>
+        <require_module>org.bahmni.module.bahmnicommons</require_module>
     </require_modules>
     <!-- Extensions -->
