+++ pom.xml
@@ -48,6 +48,7 @@
         <idgenWebServicesVersion>1.3-SNAPSHOT</idgenWebServicesVersion>
         <providermanagementVersion>2.13.0</providermanagementVersion>
         <velocityVersion>1.6.2</velocityVersion>
+        <bahmniCommons.version>0.2-SNAPSHOT</bahmniCommons.version>
 
         <!--test dependencies-->
         <junitVersion>4.13</junitVersion>
@@ -188,6 +189,13 @@
                         <artifactId>servlet-api</artifactId>
                     </exclusion>
                 </exclusions>
+                <scope>provided</scope>
+            </dependency>
+            <dependency>
+                <groupId>org.bahmni.module</groupId>
+                <artifactId>bahmni-commons-api</artifactId>
+                <type>jar</type>
+                <version>${bahmniCommons.version}</version>
                 <scope>provided</scope>
             </dependency>
             <dependency>
